requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
dynamsoft_capture_vision_bundle==3.2.5000
flatbuffers==25.12.19
fonttools==4.61.1
jax==0.6.2
jaxlib==0.6.2
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.14
ml_dtypes==0.5.4
mrz-scanner-sdk==4.0.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==4.25.8
pycparser==2.23
pycryptodome==3.23.0
PyMuPDF==1.26.7
pyparsing==3.3.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzbar==0.1.9
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.3
threadpoolctl==3.6.0
tzdata==2025.3
zxing-cpp==2.3.0
